requirements.txt
playwright>=1.44.0
httpx>=0.27.0
beautifulsoup4>=4.12.3
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic[email]>=2.10.0
schedule>=1.2.1
python-dotenv>=1.0.1
